[PATCH] Add testing and coverage ci
