+++ setup.cfg
@@ -3,12 +3,8 @@ exclude = docs/*,config/*,.github/*
 ignore = E123,E128,E402,W503,W504,E731,W601
 max-line-length = 120
 
-[nosetests]
-cover-package=ftlengine
-cover-erase=
-
 [tool:pytest]
-addopts = --cov=ftlengine --cov-report term-missing:skip-covered --cov-config=setup.cfg
+addopts = --cov --cov-report term-missing:skip-covered --cov-config=setup.cfg
 
 [coverage:run]
 omit = 
@@ -18,6 +14,7 @@ omit =
     /usr/*
     # omit tests folder
     tests/*
+    .eggs/*
 
 [coverage:report]
 exclude_lines =
